// rotv-flight/src/gen/rotv/v1/common_pb.d.ts
export enum CheckStatus {
  /**
   * @generated from enum value: CHECK_STATUS_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: CHECK_STATUS_PENDING = 1;
   */
  PENDING = 1,

  /**
   * @generated from enum value: CHECK_STATUS_RUNNING = 2;
   */
  RUNNING = 2,

  /**
   * @generated from enum value: CHECK_STATUS_PASSED = 3;
   */
  PASSED = 3,

  /**
   * @generated from enum value: CHECK_STATUS_FAILED = 4;
   */
  FAILED = 4,

  /**
   * @generated from enum value: CHECK_STATUS_SKIPPED = 5;
   */
  SKIPPED = 5,

  /**
   * @generated from enum value: CHECK_STATUS_WARNING = 6;
   */
  WARNING = 6,
}